requirements.txt
telethon
python-dotenv
loguru